[No JIRA] Move the get menu for staff to store service
