+++ Pos-System/Constants/ApiEndPointConstant.cs
@@ -31,6 +31,7 @@ public static class ApiEndPointConstant
         public const string StoreEndpoint = StoresEndpoint + "/{id}";
         public const string StoreUpdateEmployeeEndpoint = StoresEndpoint + "/{storeId}/users/{id}";
         public const string StoreAccountEndpoint = StoresEndpoint + "/{storeId}/users";
+        public const string MenuProductsForStaffEndPoint = StoresEndpoint + "/menus/products";
     }
 
     public static class Account
@@ -66,6 +67,5 @@ public static class ApiEndPointConstant
         public const string MenusInBrandEndPoint = Brand.BrandEndpoint + "/menus";
         public const string HasBaseMenuEndPoint = MenusInBrandEndPoint + "/hasBaseMenu";
         public const string MenuProductsEndpoint = MenusEndPoint + "/{menuId}/products";
-        public const string MenuProductsForStaffEndPoint = MenusEndPoint + "/staff/products";
     }
 }
