+++ Pos-System/Services/Interfaces/IMenuService.cs
@@ -12,7 +12,6 @@ namespace Pos_System.API.Services.Interfaces
         public Task<HasBaseMenuResponse> CheckHasBaseMenuInBrand(Guid brandId);
         public Task<IPaginate<GetMenuDetailResponse>> GetMenus(Guid brandId, string? code, int page = 1, int size = 10);
         public Task<Guid> UpdateMenuProducts(Guid menuId, UpdateMenuProductsRequest updateMenuProductsRequest);
-        public Task<GetMenuDetailForStaffResponse> GetMenuDetailForStaff();
         public Task<GetMenuDetailResponse> GetMenuDetailInBrand(Guid menuId);
         public Task<IPaginate<GetProductInMenuResponse>> GetProductsInMenu(Guid menuId, string? productName, int page, int size);
     }
