+++ Pos-System/Services/Interfaces/IStoreService.cs
@@ -1,4 +1,5 @@
 ﻿using Pos_System.API.Payload.Request.Stores;
+using Pos_System.API.Payload.Response.Menus;
 using Pos_System.API.Payload.Response.Stores;
 using Pos_System.Domain.Paginate;
 
@@ -15,5 +16,6 @@ public interface IStoreService
     public Task<IPaginate<GetStoreEmployeesResponse>> GetStoreEmployees(Guid storeId, string? searchUserName, int page, int size);
 
     public Task<bool> UpdateStoreInformation(Guid storeId, UpdateStoreRequest updateStoreRequest);
+    public Task<GetMenuDetailForStaffResponse> GetMenuDetailForStaff();
 
 }
